[78] define a Higher-Order function num1AndNum2:  If a function takes another function as param, or if the return type is another function, then this function is a higher-order function.
In num1AndNum2(), the first two params are of type Int, the third param is a function type param that takes two integer params and return value of integer type.

In num1AndNum2(), we do not specify any operations. Instead, we pass num1 and num2 to the function type param and return the return value of the function type param.

Since num1AndNum2() takes a function type param, thus we need to define the corresponding function. We define two functions that match the param list and also the return value of the function type param in num1AndNum2().

The plus() will add the two integers together and return the sum, and minus() will calculate the difference and return.

The third argument uses ::plus and ::minus syntax. This is a way to reference the functions which means passing plus() and minus() as an argument to num1AndNum2().