[81]  add inline modifier to our Higher-Order Function for performance optimization:
What is happening behind the scenes of Kotlin's higher-order function? The Lambda expression is being turned to anonymous class under the hood at runtime. This means that every time we call Lambda expression, an instance of anonymous class will be created which leads to extra cost on memory and performance.

In order to solve this problem, Kotlin provides inline function which can eliminate the runtime cost of Lambda expressions.

Kotlin compiler will put the code inside inline functions to the place that is calling it, which means there is no extra cost at run time.