[88] create an example of delegation properties: the "by" keyword connects the p field and the Delegate instance.  This means that p field’s implementation is delegated to Delegate class. When p field gets read, then getValue() in Delegate will be called, and when p field gets written, setValue() in Delegate will be called.
In Delegate class, we need to implement getValue() and setValue() and use the "operator" keyword to declare.

* getValue() takes two params: the first param is used to specify which class can delegate to this Delegate class, and here we set it to be MyClass; the second param KProperty<*> is a property operator class in Kotlin, which can be used to get the values related to property and is not used here but is required to be in the method param list. <*> means that you don’t know or don’t care for the type of the generics and just want to compile successfully which is like <?> in Java. The return value can be any type based on the implementation logic, the above code is just for demonstration purpose.

* setValue() is similar except that it takes three params. The first two params are the  same as in getValue(), the last param is the value that will be assigned to  delegation property which should be the same type of the return value of  getValue().

The delegation property work flow is like this: when we assign value to p property  of MyClass, then setValue() of Delegate will be called, and when we try to get the  value of p property, then getValue() of Delegate will be called.

If p field is declared as val then we do not need to implement setValue(). This  should be easy to understand as if p is a val, then it cannot be reassigned after  initialization, thus there is no need to implement setValue(), we only need to  implement getValue().